Revert "Merge branch 'master' into rgbMatrix-screenGrabber"
This reverts commit 23232528704cd463611ae6a4cb7cdc40d0e8ed65, reversing
changes made to 728e3f676e1ad9a47f28227fdb0f880da68a922b.

+++ engine/src/rgbscript.cpp
@@ -36,6 +36,8 @@
 
 #include "rgbscript.h"
 #include "rgbscriptscache.h"
+#include "qlcconfig.h"
+#include "qlcfile.h"
 
 QScriptEngine* RGBScript::s_engine = NULL;
 #if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
@@ -61,7 +63,7 @@ RGBScript::RGBScript(const RGBScript& s)
     , m_apiVersion(0)
 {
     evaluate();
-    foreach (RGBScriptProperty cap, s.m_properties)
+    foreach(RGBScriptProperty cap, s.m_properties)
     {
         setProperty(cap.m_name, s.property(cap.m_name));
     }
@@ -79,7 +81,7 @@ RGBScript &RGBScript::operator=(const RGBScript &s)
         m_contents = s.m_contents;
         m_apiVersion = s.m_apiVersion;
         evaluate();
-        foreach (RGBScriptProperty cap, s.m_properties)
+        foreach(RGBScriptProperty cap, s.m_properties)
         {
             setProperty(cap.m_name, s.property(cap.m_name));
         }
@@ -366,7 +368,7 @@ QHash<QString, QString> RGBScript::propertiesAsStrings()
     QMutexLocker engineLocker(s_engineMutex);
 
     QHash<QString, QString> properties;
-    foreach (RGBScriptProperty cap, m_properties)
+    foreach(RGBScriptProperty cap, m_properties)
     {
         QScriptValue readMethod = m_script.property(cap.m_readMethod);
         if (readMethod.isFunction())
@@ -390,7 +392,7 @@ bool RGBScript::setProperty(QString propertyName, QString value)
 {
     QMutexLocker engineLocker(s_engineMutex);
 
-    foreach (RGBScriptProperty cap, m_properties)
+    foreach(RGBScriptProperty cap, m_properties)
     {
         if (cap.m_name == propertyName)
         {
@@ -421,7 +423,7 @@ QString RGBScript::property(QString propertyName) const
 {
     QMutexLocker engineLocker(s_engineMutex);
 
-    foreach (RGBScriptProperty cap, m_properties)
+    foreach(RGBScriptProperty cap, m_properties)
     {
         if (cap.m_name == propertyName)
         {
@@ -476,7 +478,7 @@ bool RGBScript::loadProperties()
         RGBScriptProperty newCap;
 
         QStringList propsList = cap.split("|");
-        foreach (QString prop, propsList)
+        foreach(QString prop, propsList)
         {
             QStringList keyValue = prop.split(":");
             if (keyValue.length() < 2)
@@ -493,7 +495,7 @@ bool RGBScript::loadProperties()
             else if (key == "type")
             {
                 if (value == "list") newCap.m_type = RGBScriptProperty::List;
-                else if (value == "float") newCap.m_type = RGBScriptProperty::Float;
+                else if (value == "integer") newCap.m_type = RGBScriptProperty::Integer;
                 else if (value == "range") newCap.m_type = RGBScriptProperty::Range;
                 else if (value == "string") newCap.m_type = RGBScriptProperty::String;
             }
